GetLocation Moved to APICalls
