+++ .expo/packager-info.json
@@ -2,9 +2,9 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19003,
-  "packagerPid": 89588,
+  "packagerPid": 95733,
   "expoServerNgrokUrl": "https://xr-4p8.anonymous.mvp-map.exp.direct",
   "packagerNgrokUrl": "https://packager.xr-4p8.anonymous.mvp-map.exp.direct",
-  "ngrokPid": 89618,
-  "webpackServerPort": null
+  "ngrokPid": 95763,
+  "webpackServerPort": 19006
 }
