+++ screens/Main.js
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
 import { StyleSheet, Text, View, } from 'react-native';
-import * as Permissions from 'expo-permissions';
-import * as Location from 'expo-location';
-import { useSelector, useDispatch } from 'react-redux';
+
+import { useSelector, } from 'react-redux';
 
 import WelcomeView from './WelcomeView';
 import MapViewScreen from './MapViewScreen';
@@ -10,37 +9,15 @@ import ActivityView from './ActivityView';
 import SettingsView from './SettingsView';
 
 import { colors } from '../Constants/Colors';
-import { setLocation, setCurrentLocation } from '../store/actions/locationActions';
+import { getLocation, } from '../Constants/APICalls';
 
 const Main = () => {
   const viewMode = useSelector((state) => state.screen.screen);
   const location = useSelector((state) => state.location.location);
-  const current = useSelector((state) => state.location.current);
-
-  const dispatch = useDispatch();
-
-  // REVIEW
-  const getLocation = async () => {
-    try {
-      let { status } = await Permissions.askAsync(Permissions.LOCATION);
-      if (status !== 'granted') {
-        Location.requestPermissionsAsync();
-        dispatch(setLocation(current))
-      } else {
-        let location = await Location.getCurrentPositionAsync({timeout: 10000});
-        dispatch(setLocation(location));
-        dispatch(setCurrentLocation(location));
-      }
-    } catch {
-      console.log('error in getLocation ');
-    }
-  };
   
   useEffect(() => {
-    if (location === null){
-      getLocation();
-    }
-  }, [location]);
+    getLocation();
+  }, []);
 
   let screen;
   switch(viewMode) {
